refact: move to_ methods from VAR to ID_
