+++ src/symbols/var.py
@@ -67,41 +67,6 @@ class SymbolVAR(SymbolID):
 
         return "$" + self._t  # Local string variables (and parameters) use '$' (see backend)
 
-    @staticmethod
-    def to_label(var_instance):
-        """Converts a var_instance to a label one"""
-        # This can be done 'cause LABEL is just a dummy descent of VAR
-        assert isinstance(var_instance, SymbolVAR)
-        from src.symbols import LABEL
-
-        var_instance.__class__ = LABEL
-        var_instance.class_ = CLASS.label
-        var_instance._scope_owner = []
-        return var_instance
-
-    @staticmethod
-    def to_function(var_instance, lineno=None, class_=CLASS.function):
-        """Converts a var_instance to a function one"""
-        assert isinstance(var_instance, SymbolVAR)
-        from src.symbols import FUNCTION
-
-        var_instance.__class__ = FUNCTION
-        var_instance.class_ = class_
-        var_instance.reset(lineno=lineno)
-        return var_instance
-
-    @staticmethod
-    def to_vararray(var_instance, bounds):
-        """Converts a var_instance to a var array one"""
-        assert isinstance(var_instance, SymbolVAR)
-        from src.symbols import BOUNDLIST, VARARRAY
-
-        assert isinstance(bounds, BOUNDLIST)
-        var_instance.__class__ = VARARRAY
-        var_instance.class_ = CLASS.array
-        var_instance.bounds = bounds
-        return var_instance
-
     @property
     def value(self):
         """An alias of default value, only available is class_ is CONST"""
